Bottom sheet auto hide/show stub implementation.
Unfortunately for now this feature will be unavailable,
because it causes issues with the BottomSheet.
The problems are related with syncing in the background
and showing the pickers in the previously opened BottomSheet.

On the other hand I have introduced two wrapper methods to quickly
show/hide the bottom sheet.